Address PR comments + exception handling for factory calls
