+++ src/test/java/jenkins/plugins/slack/StandardSlackServiceStub.java
@@ -4,10 +4,6 @@ public class StandardSlackServiceStub extends StandardSlackService {
 
     private CloseableHttpClientStub httpClientStub;
 
-    public StandardSlackServiceStub(String baseUrl, String teamDomain, String token, String tokenCredentialId, boolean botUser, String roomId) {
-        super(baseUrl, teamDomain, token, tokenCredentialId, botUser, roomId, false);
-    }
-
     public StandardSlackServiceStub(String baseUrl, String teamDomain, boolean botUser, String roomId, String populatedToken) {
         super(baseUrl, teamDomain, botUser, roomId, false, populatedToken);
     }
